Check TXT/SPF records for RFC1035 sec. 5.1
If record data has empty spaces it should be surrounded by double quotes.
This patch will raise an error during validation if record has whitespaces,
empty spaces, tabs, etc., but not wrapped in " " (double quotes).

Corresponding RFC part:

<character-string> is expressed in one or two ways: as a contiguous set
of characters without interior spaces, or as a string beginning with a "
and ending with a ". Inside a " delimited string any character can
occur, except for a " itself, which must be quoted using \ (back slash).

Closes-Bug: 1755788

Change-Id: I159d0732688ddf1337ab3602a84a43fd043dcaa2