Fix zone update when adding new Bind9 target to pool.
When a new Bind9 is added to the pool, Designate will trigger 'modzone'
rndc command that will fail, unless zone is already present on the backend.

This change will first verify that the zone is present on the backend,
before an update attempt. If zone is not present - zone creation will be
triggered.

Closes-Bug: 1958409
(cherry picked from commit a25634a59cc16ae969654f76a7c5489269fa4a41)