Validate matching double quotes in TXT recordsets.
Currently, designate does not validate that if a TXT recordset starts
with a double quote, it also ends with a double quote. So, if user has
provided TXT record in this unmatched doublequote format, zone ends up
in error.

This change verifies that if a TXT record starts with a double quote,
it also ends with a double quote.

Closes-Bug: 1980757
Change-Id: I75bd7c1657a9178c8e7ef2d66fa7318255c67582
(cherry picked from commit 7f428e8ab581d2a79464dd11029c397bc68d4953)