Fix update of records in the shared zone.
In case of update of a record in the shared zone we need to evaluate the
target agains the right project. If the zone is not shared we use the id
of the project it belongs to. If the zone is shared we use the id of the
current project.