Fix shared zones and transfers.
If the zone is currently shared with one or multiple projects - we need
to update 'shared_zones' when zone is transferred away from old 'parent'
project to a new 'parent' project. This way the sharing information is
transferred together with the zone itself to a new project.

During zone transfer of the shared zone we should only update tenant
ownership information for records and recordsets that were created in
the tenant from where the zone transfer is originating. Otherwise, the
records from shared projects will appear in the project where the zone
was transferred to.

Finally, the transfer of the zone into the project where it is already
shared to should not be allowed.