Fix unsharing of zones shared to multiple projects.
When we unshare a zone from parent project we should make sure that
the zone in the target_project does not have any recordsets left.

But zone can still have records in other project where it is shared to.
Therefore, we should explicitely look for records inside the
'target_project_id' project and not records excluding parent/owning project.